// puniswap-ui/src/types/token.ts
export interface Token {
  TokenID: string;
  Name: string;
  Symbol: string;
  Image: string;
  IsPrivacy: boolean;
  IsBridge: boolean;
  ExternalID: string;
  PDecimals: number;
  Decimals: number;
  ContractID: string;
  Status: number;
  Type: number;
  CurrencyType: number;
  Default: boolean;
  Verified: boolean;
  UserID: number;
  ListUnifiedToken: Token[];
  PSymbol: string;
  OriginalSymbol: string;
  LiquidityReward: number;
  ExternalPriceUSD: number;
  PriceUsd: number;
  PercentChange1h: string;
  PercentChangePrv1h: string;
  PercentChange24h: string;
  CurrentPrvPool: number;
  PricePrv: number;
  volume24: number;
  ParentID: number;
  Network: string;
  DefaultPoolPair: string;
  DefaultPairToken: string;
  NetworkID: number;
  MovedUnifiedToken: boolean;
  ParentUnifiedID: number;
  IsSwapable: boolean;
  ContractIDSwap: string;
}